+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Updated if statements
 - Changed access syntax
 - Updated import statements
+- Added static blocks
 
 ### [Nishi-0.12.6-alpha](https://github.com/DeflatedPickle/Nishi/releases/tag/v0.12.6-alpha)
 - Added an argument for pretty printing